windows: Fix compilation issue
This commit is to make sure there is no compilication issue
with windows.

Fixes #146
