+++ xfrm_unspecified.go
@@ -1,0 +1,10 @@
+//go:build !linux
+// +build !linux
+
+package netlink
+
+import "fmt"
+
+func (p Proto) String() string {
+	return fmt.Sprintf("%d", p)
+}
